rename initPlugins => core
