+++ src/core.js
