+++ src/index.js
@@ -1,7 +1,7 @@
 import init from './init'
 import model from './model'
 import { getStore } from './store'
-import { pluginExports } from './initPlugins'
+import { pluginExports } from './core'
 
 export default {
   init,
