+++ src/model.js
@@ -1,7 +1,7 @@
 // @flow
 import validate from './utils/validate'
 import { createReducersAndUpdateStore, getStore } from './store'
-import { onModelHooks, pluginExports } from './initPlugins'
+import { onModelHooks, pluginExports } from './core'
 
 const validateModel = (model: $model) =>
   validate([
